requirements.txt
annotated-types==0.7.0
anyio==4.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==2.3.1
Babel==2.8.0
backcall==0.2.0
bandit==1.7.10
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
brotlicffi==1.1.0.0
cachetools==5.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
ddt==1.7.2
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.115.3
fastapi-cli==0.0.5
fasteners==0.14.1
filelock==3.15.4
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.43
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
httptools==0.6.1
idna==3.10
importlib_metadata==8.5.0
inflect==2.1.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
meld==3.20.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.2
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.12.0
pycparser==2.22
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
PyDrive==1.3.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==22.3.0
reportlab==3.6.8
requests==2.32.3
rich==13.8.1
rsa==4.8
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
shellingham==1.5.4
six==1.16.0
smmap==6.0.0
sniffio==1.3.1
soupsieve==2.3.1
sqlacodegen==1.1.6
SQLAlchemy==2.0.34
sqlmodel==0.0.22
starlette==0.40.0
stevedore==5.3.0
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typer==0.12.5
typing-inspection==0.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==3.0.1
urllib3==2.2.3
usb-creator==0.3.7
uvicorn==0.32.0
uvloop==0.21.0
vboxapi==1.0
virtualenv==20.26.3
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==13.1
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
zipp==3.20.2
